requirements.txt
# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Deep Learning
torch>=2.0.0
tensorflow>=2.10.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.28.0
gym>=0.26.0

# Financial Data
yfinance>=0.2.18
ccxt>=4.0.0

# Optimization
cvxpy>=1.3.0
cvxopt>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.11.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0
pyyaml>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.19.0
ipywidgets>=8.0.0

# Development
pytest>=7.2.0
black>=22.0.0
flake8>=5.0.0
